integer when empty output and url update

+++ R/ffmanova_div.R
@@ -3,7 +3,7 @@
 #' A function to simulate Matlab's \sQuote{:} operator.
 #' 
 #' \code{matlabCode(a,b)} returns \code{a:b} ('s version) unless \code{a > b},
-#' in which case it returns \code{numeric(0)}.
+#' in which case it returns \code{integer(0)}.
 #' 
 #' @param from numeric.  The start value.
 #' @param to numeric.  The end value.
@@ -16,9 +16,9 @@
 #' 
 #' identical(3:5, matlabColon(3, 5)) ## => TRUE
 #' 3:1 ## => 3 2 1
-#' matlabColon(3, 1) ## => numeric(0)
+#' matlabColon(3, 1) ## => integer(0)
 #' 
-matlabColon <- function(from, to) { if(from > to) numeric(0) else from:to }
+matlabColon <- function(from, to) { if(from > to) integer(0) else from:to }
 # Author:: Bjoern-Helge Mevik
